requirements.txt
tensorflow-datasets
plac